// skeleton/base/src/util/json-api.ts
export type JsonApiError = {
    id?: string;
    links?: {
        about?: string;
        type?: string;
    };
    status: string;
    code?: string;
    title?: string;
    detail?: string;
    source?: {
        pointer?: string;
        parameter?: string;
        header?: string;
    };
    meta?: Record<string, unknown>;
};

export type JsonApiErrorResponse = {
    errors: JsonApiError[];
    meta?: Record<string, unknown>;
};